create seededelection module
