+++ util/util.go
@@ -12,9 +12,9 @@ import (
 	"encoding/json"
 	"errors"
 	"fmt"
+	"github.com/lestrrat-go/jwx/jwk"
 	"io"
 	"io/ioutil"
-	"math/rand"
 	"net"
 	"net/http"
 	"os"
@@ -23,9 +23,6 @@ import (
 	"runtime"
 	"strconv"
 	"strings"
-	"sync"
-
-	"github.com/lestrrat-go/jwx/jwk"
 
 	"github.com/ethereum/go-ethereum/common"
 
@@ -38,7 +35,6 @@ import (
 	"github.com/chainpoint/chainpoint-core/types"
 )
 
-var randSourceLock sync.Mutex
 
 // LogError : Log error if it exists
 func LogError(err error) error {
@@ -119,28 +115,6 @@ func GetIPOnly(ip string) string {
 		listenAddr = listenAddr[:strings.LastIndex(listenAddr, ":")]
 	}
 	return listenAddr
-}
-
-// GetSeededRandInt : Given a seed and a maximum size, generates a random int between 0 and upperBound
-func GetSeededRandInt(seedBytes []byte, upperBound int) int {
-	eightByteHash := seedBytes[0:7]
-	seed, _ := binary.Varint(eightByteHash)
-	randSourceLock.Lock()
-	rand.Seed(seed)
-	index := rand.Intn(upperBound)
-	randSourceLock.Unlock()
-	return index
-}
-
-// GetSeededRandFloat : Given a seed return a random float
-func GetSeededRandFloat(seedBytes []byte) float32 {
-	eightByteHash := seedBytes[0:7]
-	seed, _ := binary.Varint(eightByteHash)
-	randSourceLock.Lock()
-	rand.Seed(seed)
-	index := rand.Float32()
-	randSourceLock.Unlock()
-	return index
 }
 
 // UUIDFromHash : generate a uuid from a byte hash, must be 16 bytes
@@ -409,91 +383,6 @@ func Contains(s interface{}, elem interface{}) bool {
 		}
 	}
 	return false
-}
-
-// Rotate rotates the values of a slice by rotate positions, preserving the
-// rotated values by wrapping them around the slice
-func rotate(slice interface{}, rotate int) error {
-	// check slice is not nil
-	if slice == nil {
-		return errors.New("non-nil slice interface required")
-	}
-	// get the slice value
-	sliceV := reflect.ValueOf(slice)
-	if sliceV.Kind() != reflect.Slice {
-		return errors.New("slice kind required")
-	}
-	// check slice value is not nil
-	if sliceV.IsNil() {
-		return errors.New("non-nil slice value required")
-	}
-
-	// slice length
-	sLen := sliceV.Len()
-	// shortcut when empty slice
-	if sLen == 0 {
-		return nil
-	}
-
-	// limit rotates via modulo
-	rotate %= sLen
-	// Go's % operator returns the remainder and thus can return negative
-	// values. More detail can be found under the `Integer operators` section
-	// here: https://golang.org/ref/spec#Arithmetic_operators
-	if rotate < 0 {
-		rotate += sLen
-	}
-	// shortcut when shift == 0
-	if rotate == 0 {
-		return nil
-	}
-
-	// get gcd to determine number of juggles
-	gcd := gcd(rotate, sLen)
-
-	// do the shifting
-	for i := 0; i < gcd; i++ {
-		// remember the first value
-		temp := reflect.ValueOf(sliceV.Index(i).Interface())
-		j := i
-
-		for {
-			k := j + rotate
-			// wrap around slice
-			if k >= sLen {
-				k -= sLen
-			}
-			// end when we're back to where we started
-			if k == i {
-				break
-			}
-			// slice[j] = slice[k]
-			sliceV.Index(j).Set(sliceV.Index(k))
-			j = k
-		}
-		// slice[j] = slice
-		sliceV.Index(j).Set(temp)
-		// elemJ.Append(temp)
-	}
-	// success
-	return nil
-}
-
-func gcd(x, y int) int {
-	for y != 0 {
-		x, y = y, x%y
-	}
-	return x
-}
-
-// RotateRight is the analogue to RotateLeft
-func RotateRight(slice interface{}, by int) error {
-	return rotate(slice, -by)
-}
-
-// RotateLeft is an alias for Rotate
-func RotateLeft(slice interface{}, by int) error {
-	return rotate(slice, by)
 }
 
 //ReadContractJSON : reads in TierionNetworkToken and ChainpointRegistry json files
